Move authentication controller to controllers folder
