+++ src/controllers/authentication/command/sign-up.command.handler.ts
