+++ src/controllers/authentication/command/sign-up.command.ts
