+++ src/domain/service/authentication/authentication.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { UserService } from '../user/user.service';
-import { UserCreationDto } from '../../../authentication/dto/user-creation.dto';
 import { User } from '../../entity/user.entity';
+import { UserCreationDto } from '../../../controllers/authentication/dto/user-creation.dto';
 
 @Injectable()
 export class AuthenticationService {
